feat(WORKSPACE)!: Updating PyTorch version to 1.8.1
BREAKING CHANGE: Due to issues with compatability between PyTorch 1.8.0
and 1.8.1 in the Torch Python API, TRTorch 0.3.0 compiled for 1.8.0 does not
work with PyTorch 1.8.1 and will show an error about use_input_stats.
If you see this error make sure the version of libtorch you are
compiling with is PyTorch 1.8.1

TRTorch 0.3.0 will target PyTorch 1.8.1. There is no backwards
compatability with 1.8.0. If you need this specific version compile from
source with the dependencies in WORKSPACE changed

Signed-off-by: Naren Dasan <naren@narendasan.com>
Signed-off-by: Naren Dasan <narens@nvidia.com>