refactor: removing the strict_types and max_batch_size apis
BREAKING CHANGE: This commit removes the strict types and max_batch_size apis. We are doing this because the functionality of these APIs in TRT is convoluted and likely to be ignored during building. A replacement for strict types with actual guarantees will be added at a later date.

For max_batch_size, this argument has always been ignored since we use explicit batch sizing. To set the batch size, use the first dimension of your input type

Signed-off-by: Dheeraj Peri <peri.dheeraj@gmail.com>