refactor(//core/conversion/var): Modify var contract
This modifies the variable contract to allow requesting of
mutable handles to `IValue`s which previously were provided
only as `const`. Now every IValue pointer is not `const` within
`Var`, but in order to get a `const` handle you use IValueMut.

This supports the evaluation of inplace operations and makes
state changes explicit rather than implicit.

Signed-off-by: Naren Dasan <naren@narendasan.com>
Signed-off-by: Naren Dasan <narens@nvidia.com>