fix: Allow full model compilation with collection outputs
- Update graph-building in compiler to account for case where all
operations are supported by Torch-TRT, but the output is a collection.
- Enable 'psuedo-partitioning' for nearly-fully-compiled models for
which the only non-supported aspect of the model is the format of the
output (TRT cannot output complex collections)
- Define a small subset of operation schemas which are allowed despite
the flag `require_full_compilation`. These operations are packing and
unpacking of Tuples/Lists, and some are already used in cases of
`require_full_compilation`
- Display warnings to users if any portion of the `pseudo-partitioning`
is unexpected, for example the model being partitioned ends up in more
than 3 segments (maximally - a Torch segment to preprocess collection
inputs, a TRT segment to perform model logic, a Torch segment to
post-process collection outputs) or if schemas falling outside of the
collection subset are encountered in a Torch segment
- Add end-to-end test case with minimal reproducing example of a failing
model, repaired with the changes to the compiler
- Add minor fix to lowering to remediate c++ compiler warning