Disambiguate element-wise cast layer names
Duplicate layer names in TRT cause an error. Cast layers by default are named based on the Tensor pointer initial type and target type. This could cause name conflicts with multiple element-wise operations that performed the same type promotion on a given tensor. To resolve this, add the element_wise layer name to the type-promotion cast layer as a prefix.